requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
python-multipart
passlib
requests
langchain
langchain-core
langchain-openai
langchain-chroma
langserve
langgraph
sse-starlette
ipykernel # for running Jupyter Notebook cells